Creating a libparanumal operator
