+++ Makefile
@@ -34,9 +34,6 @@ ifneq ($(wildcard ../mfem/libmfem.*),)
 endif
 
 # OCCA_DIR env variable should point to OCCA master (github.com/libocca/occa)
-LIBP_DIR ?= ../libparanumal
-
-# OCCA_DIR env variable should point to OCCA master (github.com/libocca/occa)
 OCCA_DIR ?= ../occa
 
 # env variable MAGMA_DIR can be used too
@@ -214,6 +211,10 @@ ifneq ($(wildcard $(OCCA_DIR)/lib/libocca.*),)
   libceed.c += $(occa.c)
   $(occa.c:%.c=$(OBJDIR)/%.o) : CFLAGS += -I$(OCCA_DIR)/include
   BACKENDS += /cpu/occa /gpu/occa /omp/occa
+
+  libceed.c += $(libparanumal.c)
+  $(libparanumal.c:%.c=$(OBJDIR)/%.o) : CFLAGS += -I$(OCCA_DIR)/include -Ibackends/occa
+  BACKENDS += /gpu/libparanumal
 endif
 
 ifneq ($(wildcard $(MAGMA_DIR)/lib/libmagma.*),)
@@ -233,14 +234,6 @@ ifneq ($(wildcard $(MAGMA_DIR)/lib/libmagma.*),)
   $(magma_allsrc.cu:%.cu=$(OBJDIR)/%.o) : NVCCFLAGS += --compiler-options=-fPIC -DADD_ -I$(MAGMA_DIR)/include -I$(MAGMA_DIR)/magmablas -I$(MAGMA_DIR)/control -I$(CUDA_DIR)/include
   BACKENDS += /gpu/magma
   endif
-endif
-
-ifneq ($(wildcard $(LIBP_DIR)/libP.*),)
-  $(libceed) : LDFLAGS += -L$(OCCA_DIR)/lib -Wl,-rpath,$(abspath $(OCCA_DIR)/lib)
-  $(libceed) : LDLIBS += -locca
-  libceed.c += $(libparanumal.c)
-  $(libparanumal.c:%.c=$(OBJDIR)/%.o) : CFLAGS += -I$(OCCA_DIR)/include
-  BACKENDS += /gpu/libparanumal
 endif
 
 export BACKENDS
